Moved modal service to services folder
