+++ src/app/modal/modal.component.ts
@@ -1,7 +1,7 @@
   
 ﻿import { Component, ViewEncapsulation, ElementRef, Input, OnInit, OnDestroy } from '@angular/core';
 
-import { ModalService } from './modal.service';
+import { ModalService } from '../services/modal.service';
 
 @Component({ 
     selector: 'jw-modal',
